Add files via upload
I started coding some interesting variables for the presentations. It does not follow any specific systematic. The selection is based on how interesting the information might be for the presentation in January.

Can someone ask the LimeGroup: 

I wanted to create a variable with number of children in the hh. I came across two variables:
* E_3__How_many_people_live_in_yo0 
* E_3__How_many_people_live_in_you

Which one should we take and how do I infer the # of children? Thx :)